projektZaliczeniowy/src/requirements.txt
pandas
matplotlib.pyplot
matplotlib.figure
matplotlib.backends.backend_tkagg
seaborn
tkinter
unittest